+++ tests/units/lib/ContextForTest.php
@@ -8,11 +8,13 @@
 namespace Jelix\DaoTests;
 
 use Jelix\Dao\ContextInterface;
+use Jelix\Dao\ContextInterface2;
 use Jelix\Database\AccessParameters;
 use Jelix\Database\Connection;
 use Jelix\Database\ConnectionInterface;
+use Jelix\Database\Schema\SQLSyntaxHelpersInterface;
 
-class ContextForTest implements ContextInterface
+class ContextForTest implements ContextInterface, ContextInterface2
 {
     /**
      * @var \Jelix\Database\ConnectionInterface
@@ -26,6 +28,16 @@ class ContextForTest implements ContextInterface
 
 
     protected $checkCompiledCache = true;
+
+    /**
+     * SQL type
+     */
+    protected $sqlType;
+
+    /**
+     * @var SQLSyntaxHelpersInterface
+     */
+    protected $syntaxHelpers;
 
     /**
      * ContextTest constructor.
@@ -42,7 +54,6 @@ class ContextForTest implements ContextInterface
                 'password'=>'jelixpass',
                 'database'=> 'jelixtests'
             );
-            $toolsClass = '\Jelix\Database\Schema\Mysql\SQLTools';
         }
         else if ($databaseType == 'pgsql') {
             $parameters = array(
@@ -53,7 +64,6 @@ class ContextForTest implements ContextInterface
                 'password'=>'jelixpass',
                 'database'=> 'jelixtests'
             );
-            $toolsClass = '\Jelix\Database\Schema\Postgresql\SQLTools';
         }
         else if ($databaseType == 'sqlite')
         {
@@ -61,15 +71,18 @@ class ContextForTest implements ContextInterface
                 'driver'=>'sqlite3',
                 "database"=>"/app/tests/units/tests.sqlite3",
             );
-            $toolsClass = '\Jelix\Database\Schema\Sqlite\SQLTools';
         }
         else {
             throw new \Exception('bad databaseType');
         }
 
+        $this->sqlType = $databaseType;
+
         $accessParameters = new AccessParameters($parameters, array('charset'=>'UTF-8'));
         $this->connection = Connection::create($accessParameters);
-        $this->dbTools = new $toolsClass($this->connection);
+        $this->dbTools = $this->connection->tools();
+
+        $this->syntaxHelpers = Connection::getSqlSyntaxHelpers($this->sqlType);
 
         $this->checkCompiledCache = $checkCompiledCache;
     }
@@ -85,6 +98,16 @@ class ContextForTest implements ContextInterface
     public function getDbTools()
     {
         return $this->dbTools;
+    }
+
+    public function getSqlType() : string
+    {
+        return $this->sqlType;
+    }
+
+    public function getSqlSyntaxHelpers() : SQLSyntaxHelpersInterface
+    {
+        return $this->syntaxHelpers;
     }
 
     public function resolveDaoPath($path)